+++ src/GlobalFunction.tsx
@@ -742,8 +742,14 @@ export const runValidation = (dataKey: string, updatedRef: any, activeComponentP
         if (updatedRef.type == 31 && updatedRef.answer !== undefined && typeof updatedRef.answer !== 'object') {
             let re = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
             if (!re.test(updatedRef.answer)) {
-
                 updatedRef.validationMessage.push(locale.details.language[0].validationEmail);
+                updatedRef.validationState = 2;
+            }
+        }        
+        if (updatedRef.type == 19 && updatedRef.answer !== undefined && typeof updatedRef.answer !== 'object') {
+            let re = /^https?:\/\/(?:www\.)?[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b(?:[-a-zA-Z0-9()@:%_\+.~#?&\/=]*)$/;
+            if (!re.test(updatedRef.answer)) {
+                updatedRef.validationMessage.push(locale.details.language[0].validationUrl);
                 updatedRef.validationState = 2;
             }
         }
@@ -839,7 +845,13 @@ export const saveAnswer = (dataKey: string, attributeParam: any, answer: any, ac
         addHistory('saveAnswer', dataKey, refPosition, attributeParam, reference.details[refPosition][attributeParam])
         setReference('details', refPosition, attributeParam, answer);
         //validate for its own dataKey 
-        if (referenceHistoryEnable() && (reference.details[refPosition].validations !== undefined || reference.details[refPosition].rangeInput !== undefined || reference.details[refPosition].lengthInput !== undefined || reference.details[refPosition].type == 31) && initial == 0) runValidation(dataKey, JSON.parse(JSON.stringify(reference.details[refPosition])), activeComponentPosition, prop?.clientMode);
+        if (referenceHistoryEnable() 
+            && (reference.details[refPosition].validations !== undefined 
+                || reference.details[refPosition].rangeInput !== undefined 
+                || reference.details[refPosition].lengthInput !== undefined 
+                || reference.details[refPosition].type == 31 || 19) 
+            && initial == 0) 
+        runValidation(dataKey, JSON.parse(JSON.stringify(reference.details[refPosition])), activeComponentPosition, prop?.clientMode);
 
         //do nothing if no changes, thanks to Budi's idea on pull request #5
         if (attributeParam === 'answer') {